pointerevents/pointerevent_svg-test_touch.html

<!doctype html>
<html>
    <head>
        <title>SVG test</title>
        <meta name="viewport" content="width=device-width">
        <link rel="stylesheet" type="text/css" href="pointerevent_styles.css">
        <script src="http://w3c-test.org/resources/testharness.js"></script>
        <script src="http://w3c-test.org/resources/testharnessreport.js"></script>
        <script src="pointerevent_support.js"></script>
        <style>
            #target0 {
            height: 350px;
            width: 300px;
            overflow-y: auto;
            background: black;
            padding: 100px;
            position: relative;
            }
        </style>
    </head>
    <body onload="run()">
        <h2>Pointer Events touch-action attribute support</h2>
        <h4 id="desc">Test Description: Try to scroll black element DOWN moving your touch outside of the red border. Wait for description update.</h4>
        <p>Note: this test is for touch only</p>
        <div id="target0">
            <svg width="555" height="555" style="touch-action: none;  border: 4px double red;">
                <circle cx="305" cy="305" r="250" stroke="green" stroke-width="4" fill="yellow" />
                Sorry, your browser does not support inline SVG.
            </svg>
        </div>
        <script type='text/javascript'>
            var detected_pointertypes = {};
            var xScrollIsReceived = false;
            var yScrollIsReceived = false;
            var failScrollIsReceived = false;
            var xScr0, yScr0, xScr1, yScr1;
            var scrollReturnInterval = 1000;
            var isFirstPart = true;
            var moveCount = 0;
            var countToPass = 50;
            var isFirstScroll = true;
            add_completion_callback(showPointerTypes);

            function run() {
                var target0 = document.getElementById("target0");
                var test_touchaction_div = async_test("touch-action attribute test out of SVG");
                var test_touchaction_svg = async_test("touch-action attribute test in SVG");

                xScr0 = target0.scrollLeft;
                yScr0 = target0.scrollTop;
                target0.focus();

                // catch pointer type for the cancelled touch-actions
                on_event(target0, 'pointercancel', function(event) {
                    detected_pointertypes[event.pointerType] = true;
                });

                on_event(target0, 'pointermove', function(event) {
                    detected_pointertypes[event.pointerType] = true;
                    if(!isFirstPart) {
                        moveCount++;
                        if(moveCount >= countToPass) {
                            updateDescriptionFouthStepSVG();
                            if(!isFirstScroll) {
                                test_touchaction_svg.done();
                                updateDescriptionComplete();
                            }
                            moveCount = 0;
                            isFirstScroll = false;
                        }
                    }
                });

                on_event(target0, 'scroll', function(event) {
                    if(isFirstPart) {
                        xScr1 = target0.scrollLeft;
                        yScr1 = target0.scrollTop;

                        if(xScr1 != xScr0) {
                            xScrollIsReceived = true;
                        }

                        if(yScr1 != yScr0) {
                            test_touchaction_div.step(function () {
                                yScrollIsReceived = true;
                                assert_true(true, "y-scroll received.");
                            });
                            updateDescriptionSecondStepSVG();
                        }

                        if(xScrollIsReceived && yScrollIsReceived) {
                            test_touchaction_div.done();
                            updateDescriptionThirdStepSVG();
                            setTimeout(function() {
                                isFirstPart = false;
                            }, 2 * scrollReturnInterval);
                        }
                    }
                    else {
                        if(!failScrollIsReceived) {
                            test_touchaction_svg.step(failOnScroll, "scroll received while shouldn't");
                        }
                    }
                });
            }

            function updateDescriptionSecondStepSVG() {
                window.setTimeout(function() {
                objectScroller(target0, 'up', 0);}
                , scrollReturnInterval);
                document.getElementById('desc').innerHTML = "Test Description: Try to scroll element RIGHT moving your outside of the red border";
            }

            function updateDescriptionThirdStepSVG() {
                window.setTimeout(function() {
                objectScroller(target0, 'left', 0);}
                , scrollReturnInterval);
                document.getElementById('desc').innerHTML = "Test Description: Try to scroll element DOWN starting your touch inside of the circle";
            }

            function updateDescriptionFouthStepSVG() {
                document.getElementById('desc').innerHTML = "Test Description: Try to scroll element RIGHT starting your touch inside of the circle";
            }

            function objectScroller(target, direction, value) {
                if (direction == 'up') {
                    target.scrollTop = 0;
                } else if (direction == 'left') {
                    target.scrollLeft = 0;
                }
            }
        </script>
        <h1>touch-action: none</h1>
        <div id="complete-notice">
            <p>The following pointer types were detected: <span id="pointertype-log"></span>.</p>
        </div>
        <div id="log"></div>
    </body>
</html>
